cli/requirements.txt
GitPython>=3.1.40
PyGithub>=2.1.1
requests>=2.31.0
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
